pyproject.toml
[tool.poetry]
name = "ozon-matching"
version = "0.1.0"
description = ""
authors = ["mmarin <mikhail.marin@sbermarket.ru>"]
readme = "README.md"
packages = [{include = "ozon_matching"}]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.0.1"
polars = "^0.17.14"
numpy = "^1.24.3"
matplotlib = "^3.7.1"
scikit-learn = "^1.2.2"
faiss-cpu = "^1.7.4"
pyarrow = "^12.0.0"
tqdm = "^4.65.0"
lightgbm = "^3.3.5"
catboost = "^1.2"
typer-cli = "^0.0.13"
loguru = "^0.7.0"
python-levenshtein = "^0.21.0"
optuna = "^3.1.1"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
pre-commit = "^3.3.2"
flake8 = "^4.0.1"
black = "^22.6.0"
autoflake = "^1.4"
mypy = "^0.961"
isort = "^5.10.1"
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
ruff = "^0.0.269"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | build
  | dist
  | python/notebooks
  )/
'''

[tool.isort]
profile = "black"
line_length = 88
src_paths = ["src/"]


[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true


[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py310"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10
